services/file-processor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
elasticsearch==8.11.0
redis==5.0.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
python-magic==0.4.27 